requirements.txt
numpy==2.1.2
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
seaborn==0.13.2
redshift_connector==2.1.3
ipykernel==6.29.5
jupyter==1.1.1
python-dotenv==1.0.1